requirements.txt
trimesh
numpy
flask
Flask-WTF
wtforms
scipy
requests
networkx
pyinstaller==4.5
patool